remove git ignore files
